// src/api/generated/models/partnerExternalMappingMappingType.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type PartnerExternalMappingMappingType = typeof PartnerExternalMappingMappingType[keyof typeof PartnerExternalMappingMappingType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PartnerExternalMappingMappingType = {
  COUNTERPARTY_CODE: 'COUNTERPARTY_CODE',
  PRODUCT_SKU: 'PRODUCT_SKU',
} as const;